fix(common): Extract a lot to a shared file for stats calculations
